# api/swagger.yaml
openapi: "3.0.3"
info:
  title: Code Runtime API
  description: |
    A production-grade distributed code execution platform.

    ## Overview
    Code Runtime accepts source code submissions in 40+ programming languages
    (including SQLite, MySQL, PostgreSQL, and MongoDB), executes them inside
    isolated Docker containers, and returns stdout/stderr, timing, memory
    usage, and a verdict.

    ## Authentication
    Every request (except `/health` and `/auth/token`) requires a valid bearer
    token **or** an API key passed in the `X-Auth-Token` header.

    ```
    Authorization: Bearer <jwt>
    X-Auth-Token:  <api-key>
    ```

    ## Rate Limiting
    The default rate limit is **100 requests/minute** per IP (or per API key
    for authenticated clients).  Headers `X-RateLimit-Limit`,
    `X-RateLimit-Remaining`, and `X-RateLimit-Reset` are included in every
    response.

    ## Base64 Encoding
    `source_code`, `stdin`, `expected_output`, and `stdout`/`stderr` in
    responses are **base64-encoded** to avoid JSON escaping issues with binary
    or multi-line content.
  version: "1.0.0"
  contact:
    name: Code Runtime Support
    email: support@coderuntime.io
  license:
    name: MIT
servers:
  - url: http://localhost:8002
    description: Local environment
  - url: https://code-executor.revature.com/code-service
    description: QA environment
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT obtained from POST /auth/token
    APIKeyAuth:
      type: apiKey
      in: header
      name: X-Auth-Token
      description: Static API key for programmatic access

  # -------------------------------------------------------------------------
  # Schemas
  # -------------------------------------------------------------------------
  schemas:
...
    TokenRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "demo@example.com"
        password:
          type: string
          format: password
          example: "demo"
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT bearer token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        expires_at:
          type: string
          format: date-time
          example: "2026-05-11T11:00:00Z"
        token_type:
          type: string
          example: "Bearer"
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: "validation_error"
        message:
          type: string
          description: Human-readable error description
          example: "language_id is required"
        details:
          type: object
          additionalProperties: true
          description: Optional structured details (e.g. field-level validation errors)
...
security:
  - BearerAuth: []
  - APIKeyAuth: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /auth/token:
    post:
      tags: [auth]
      summary: Generate JWT token
      description: Exchange credentials for a short-lived JWT bearer token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
            example:
              email: "demo@example.com"
              password: "demo"
      responses:
        "200":
          description: Token issued successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
              example:
                access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIiwiZXhwIjoxNzQ3MDY1NjAwfQ.abc123"
                expires_at: "2026-05-11T11:00:00Z"
                token_type: "Bearer"
        "400":
          description: Missing or malformed request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "unauthorized"
                message: "invalid email or password"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # =========================================================================
  # Submissions
  # =========================================================================
...